chore: move CurrentUser into security package
